♻️ refactor:  検索文字列まわり
Issue #278
