+++ lib/presentation/pages/repo/repo_index_page.dart
@@ -9,7 +9,7 @@ import '../../../utils/logger.dart';
 import '../../components/search_app_bar.dart';
 import 'components/repo_list_view.dart';
 import 'components/repo_sort_button.dart';
-import 'components/search_repos_text_button.dart';
+import 'components/search_repos_query.dart';
 
 /// リポジトリ一覧画面
 class RepoIndexPage extends StatefulWidget {
@@ -31,7 +31,7 @@ class RepoIndexPageState extends State<RepoIndexPage> with PageRouteAware {
         controller: _scrollController,
         slivers: [
           SearchAppBar(
-            title: const SearchReposTextButton(),
+            title: const SearchReposQueryTextButton(),
             actions: const [
               RepoSortButton(),
             ],
