+++ lib/presentation/pages/repo/repo_search_page.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import '../../components/search_app_bar.dart';
 import 'components/query_histories_list_view.dart';
 import 'components/repo_sort_button.dart';
-import 'components/search_repos_text_field.dart';
+import 'components/search_repos_query.dart';
 
 /// リポジトリ検索画面
 class RepoSearchPage extends StatelessWidget {
@@ -19,7 +19,7 @@ class RepoSearchPage extends StatelessWidget {
       body: CustomScrollView(
         slivers: [
           SearchAppBar(
-            title: const SearchReposTextField(),
+            title: const SearchReposQueryTextField(),
             actions: const [
               RepoSortButton(),
             ],
